+++ Payload_Type/athena/athena/mythic/agent_functions/cp.py
@@ -1,6 +1,6 @@
 from mythic_container.MythicCommandBase import *
 from mythic_container.MythicRPC import *
-from .athena_messages import message_converter
+from .athena_utils import message_converter
 import json
