+++ Payload_Type/athena/athena/mythic/agent_functions/load-assembly.py
@@ -7,7 +7,7 @@ from mythic_container.MythicRPC import *
 from os import listdir
 from os.path import isfile, join, exists
 
-from .athena_messages import message_converter
+from .athena_utils import message_converter
 
 # create a class that extends TaskArguments class that will supply all the arguments needed for this command
 class LoadAssemblyArguments(TaskArguments):
