+++ Payload_Type/athena/athena/mythic/agent_functions/rm.py
@@ -3,7 +3,7 @@ from mythic_container.MythicCommandBase import *
 import json
 from mythic_container.MythicRPC import *
 
-from .athena_messages import message_converter
+from .athena_utils import message_converter
 
 
 class RmArguments(TaskArguments):
