+++ Payload_Type/athena/athena/mythic/agent_functions/ssh.py
@@ -3,7 +3,7 @@ import json
 import sys
 from mythic_container.MythicRPC import *
 
-from .athena_messages import message_converter
+from .athena_utils import message_converter
 
 class SshArguments(TaskArguments):
     def __init__(self, command_line, **kwargs):
