+++ Payload_Type/athena/athena/mythic/agent_functions/token.py
@@ -2,7 +2,7 @@ from mythic_container.MythicCommandBase import *
 from mythic_container.MythicRPC import *
 import json
 
-from .athena_messages import message_converter
+from .athena_utils import message_converter
 
 
 class TokenArguments(TaskArguments):
