+++ Payload_Type/athena/athena/mythic/agent_functions/upload.py
@@ -2,7 +2,7 @@ from mythic_container.MythicCommandBase import *
 from mythic_container.MythicRPC import *
 import sys
 
-from .athena_messages import message_converter
+from .athena_utils import message_converter
 
 class UploadArguments(TaskArguments):
     def __init__(self, command_line, **kwargs):
